Pure refactoring: split Platform.cs into one file per type
Arguably this demonstrates that we should have used nested classes or another namespace, but it's at least less unwieldy when split up. (In particular, it's easier to read all of Platform itself.)
